refactoring of BoardContext
- instead of exposing the dispatch function we now expose the actions
- extracted action types and reducer into its own files
