+++ frontend/src/context/BoardContext.js
@@ -1,8 +1,12 @@
 import React, { useReducer } from "react";
 import io from "socket.io-client";
 
+import { reducer } from "../reducers/boardReducer";
 import { BACKEND_ENDPOINT } from "../utils";
-import { SET_FOCUSED_CARD, REMOVE_FOCUSED_CARD } from "../actions/actionTypes";
+import {
+  SET_FOCUSED_CARD,
+  REMOVE_FOCUSED_CARD
+} from "../actionTypes/boardTypes";
 
 export const BoardContext = React.createContext();
 
@@ -10,36 +14,30 @@ export const BoardContext = React.createContext();
 // TODO: move socket logic here
 const initialState = { focusedCard: "" };
 
-function reducer(state, action) {
-  switch (action.type) {
-    case SET_FOCUSED_CARD:
-      return {
-        focusedCard: action.payload.focusedCard
-      };
-    case REMOVE_FOCUSED_CARD:
-      return {
-        focusedCard: ""
-      };
-    default:
-      return state;
-  }
-}
-
 let socket;
 
 export const BoardContextProvider = props => {
   const boardId = props.match.params.boardId;
-  const [boardState, boardDispatch] = useReducer(reducer, initialState);
+  const [boardState, dispatch] = useReducer(reducer, initialState);
 
   if (!socket) {
     socket = io(BACKEND_ENDPOINT, { query: "boardId=" + boardId });
   }
 
+  const setFocusedCard = focusedCard => {
+    dispatch({ type: SET_FOCUSED_CARD, payload: { focusedCard } });
+  };
+
+  const removeFocusedCard = () => {
+    dispatch({ type: REMOVE_FOCUSED_CARD });
+  };
+
   const value = {
     boardId,
     boardState,
-    boardDispatch,
-    socket
+    socket,
+    setFocusedCard,
+    removeFocusedCard
   };
 
   return (
